Move Yast2::Issues to Y2Issues
